requirements.txt
boto3==1.16.34
botocore==1.19.34
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
passlib==1.7.4
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
smartystreets-python-sdk==4.7.2
urllib3==1.26.2
Werkzeug==1.0.1
